added Kitab font to Surah Name
This font works really well with Quranic text. Currently, some Surah Names are breaking in the middle of the word or so because the 'Amiri' font doesn't know how to to process some shapes, this font solve the issue. The interesting thing is that the words breaking in half issue only occured in the Surah Name and not in the verses even though the verses have way more words in them so it should have occured long time ago but it didn't, in case it does in the future I will swap everything Quranic to use this font. Weird, though.